// web/src/lib/types.ts
export type FileType = "media" | "photo" | "video" | "audio" | "file";
// ...
export type Auto = {
  preload: {
    enabled: boolean;
  };
  download: {
    enabled: boolean;
    rule: AutoDownloadRule;
  };
  transfer: {
    enabled: boolean;
    rule: AutoTransferRule;
  };
};

export const TransferPolices = ["DIRECT", "GROUP_BY_CHAT", "GROUP_BY_TYPE", "GROUP_BY_AI"] as const;
export type TransferPolicy = (typeof TransferPolices)[number];
export const DuplicationPolicies = [
  "OVERWRITE",
  "RENAME",
  "SKIP",
  "HASH",
] as const;
export type DuplicationPolicy = (typeof DuplicationPolicies)[number];

export type AutoTransferRule = {
  transferHistory: boolean;
  destination: string;
  transferPolicy: TransferPolicy;
  duplicationPolicy: DuplicationPolicy;
  extra: Record<string, any>
};

export type AutoDownloadRule = {
  query: string;
  fileTypes: Array<Exclude<FileType, "media">>;
  downloadHistory: boolean;
  downloadCommentFiles: boolean;
  filterExpr: string;
};